add rule groups
