+++ src/global-messages/types.ts
@@ -1,7 +1,7 @@
 import {Statuses} from '@atb/theme';
 import {LanguageAndTextType} from '@atb/translations';
 import {FirebaseFirestoreTypes} from '@react-native-firebase/firestore';
-import {AnyMode} from '@atb/components/icon-box';
+import {Rule} from './rules';
 
 export type AppPlatformType = 'ios' | 'android';
 
@@ -33,29 +33,3 @@ export type GlobalMessageType = Omit<
   GlobalMessageRaw,
   'appPlatforms' | 'appVersionMin' | 'appVersionMax' | 'startDate' | 'endDate'
 > & {startDate?: number; endDate?: number};
-
-export type RuleVariables = {
-  modes?: AnyMode[];
-  fromZones?: string[];
-  toZones?: string[];
-  platform?: string;
-  count?: number;
-};
-
-export type Rule = {
-  variable: string; // key of RuleVariables
-  operator: RuleOperator;
-  value: string | number | boolean;
-  passOnUndefined?: boolean;
-};
-
-export enum RuleOperator {
-  equalTo = 'equalTo',
-  notEqualTo = 'notEqualTo',
-  greaterThan = 'greaterThan',
-  lessThan = 'lessThan',
-  greaterThanOrEqualTo = 'greaterThanOrEqualTo',
-  lessThanOrEqualTo = 'lessThanOrEqualTo',
-  contains = 'contains',
-  notContains = 'notContains',
-}
